Move RequestViewModel to Network package
