+++ app/src/main/java/ch/heigvd/pro/b04/android/Poll/PollViewModel.kt
@@ -3,7 +3,7 @@ package ch.heigvd.pro.b04.android.Poll
 import android.app.Application
 import androidx.lifecycle.MutableLiveData
 import ch.heigvd.pro.b04.android.Datamodel.Question
-import ch.heigvd.pro.b04.android.Utils.RequestsViewModel
+import ch.heigvd.pro.b04.android.Network.RequestsViewModel
 
 class PollViewModel(application: Application,
                     idModerator : Int,
